requirements.txt
pynput>=1.7.6
Pillow>=10.0.0 